requirements.txt
langchain>=0.1.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
# langchain-openai==0.3.8  # 不再使用OpenAI
# openai==1.65.4  # 不再使用OpenAI
openai>=1.0.0  # DeepSeek支持OpenAI SDK格式，但我们使用它来调用DeepSeek API
tavily-python==0.5.1
faiss-cpu==1.10.0
streamlit==1.43.1
sentence-transformers>=2.2.0  # 用于BGE-M3嵌入模型
langchain-huggingface>=0.0.2  # 用于HuggingFaceEmbeddings的最新版本 
zhipuai